+++ src/auth/cookie/fetch-and-persist-auth-session.ts
