+++ src/auth/kratos.ts
@@ -1,7 +1,5 @@
 import { Configuration as KratosConfig, V0alpha2Api } from '@ory/client'
 
 export const kratos = new V0alpha2Api(
-  new KratosConfig({
-    basePath: `/api/.ory`,
-  })
+  new KratosConfig({ basePath: `/api/.ory` })
 )
