+++ src/components/pages/auth/logout.tsx
@@ -2,8 +2,8 @@ import { useRouter } from 'next/router'
 import { useEffect } from 'react'
 
 import { filterUnwantedRedirection } from './utils'
-import { AuthSessionCookie } from '@/auth/auth-session-cookie'
-import { fetchAndPersistAuthSession } from '@/auth/fetch-auth-session'
+import { AuthSessionCookie } from '@/auth/cookie/auth-session-cookie'
+import { fetchAndPersistAuthSession } from '@/auth/cookie/fetch-and-persist-auth-session'
 import { kratos } from '@/auth/kratos'
 import { AxiosError } from '@/auth/types'
 import { useAuth } from '@/auth/use-auth'
